1시도 1차 평가 피드백 반영

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: seokchoi <seokchoi@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 16:26:28 by seokchoi          #+#    #+#             */
-/*   Updated: 2022/03/15 19:54:28 by seokchoi         ###   ########.fr       */
+/*   Updated: 2022/04/05 12:42:39 by seokchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
